Added handling for not existing managers in change_old_diff_widgets

+++ app/database/migrations/2015_09_29_154442_change_old_differentiated_widgets.php
@@ -30,7 +30,9 @@ class ChangeOldDifferentiatedWidgets extends Migration {
                     $old_manager = $widget->data->manager;
                     $widget->descriptor()->associate($descriptor);
                     $widget->saveSettings(array('type' => 'diff'));
-                    $old_manager->delete();
+                    if (!is_null($old_manager)) {
+                        $old_manager->delete();
+                    }
                     Log::info("Linked widget #" . $widget->id . " data to cumulative.");
                 }
